Create accreditNFT Tezos Wordpress Login
This WordPress plugin allows users to sign in to the website using their Temple wallet for Tezos. When a user clicks the "Sign in with Temple Wallet" button on the login form, the plugin opens the Temple wallet login popup. The user can then sign in to their Temple wallet and authorize the login.

Once the login is successful, the plugin sends the user's wallet address to the server, which checks if a user with this wallet address already exists. If a user with this wallet address does not exist, the plugin creates a new user with this wallet address as the username.

Finally, the plugin logs the user in by setting the current user and setting the authentication cookie. This allows the user to access the website as an authenticated user.