Update Tezos NFT Gallery Wordpress Plugin
This is a WordPress plugin that requests permission from the Temple wallet to pull NFTs owned by the user and display their images.

This plugin adds a shortcode that can be used to display a gallery of NFTs owned by the user. The shortcode adds a button to the page that, when clicked, requests permission from the Temple wallet to access the user's NFTs. If permission is granted, the plugin retrieves the user's NFTs and displays their images on the page. If permission is denied, an alert is displayed.

To use this plugin, simply add the shortcode [nft_gallery] to any page or post on your WordPress website. This will add the NFT gallery to the page, allowing the user to refresh and view their NFTs.