Create Tezos NFT Gallery for Registered Users
This WordPress plugin allows registered users to associate their NFTs with their account by entering their Tezos wallet address in their user profile. The plugin stores the wallet address in the user's profile and displays a gallery of the user's NFTs on their profile page.

The plugin adds a shortcode that can be used to display the NFT gallery on the user's profile page. The shortcode adds a button to the page that, when clicked, requests permission from the Temple wallet to access the user's NFTs. If permission is granted, the plugin retrieves the user's NFTs and displays their images on the page. If permission is denied, an alert is displayed.

In addition, the plugin adds a field to the user's profile where they can enter their Tezos wallet address. This allows the plugin to associate the user's NFTs with their account based on their wallet address. This way, other registered users can view the user's NFTs by visiting their profile page.

The plugin also allows registered users to view one another's NFTs by visiting each other's profile pages. The plugin stores the wallet address in the user's profile and displays a gallery of the user's NFTs on their profile page. Other registered users can visit a user's profile page and view their NFTs by clicking the "Refresh NFTs" button and granting permission to access their NFTs. This way, registered users can view and explore one another's NFT collections by visiting their profile pages and interacting with the NFT gallery.