Create Tezos Payments for WooCommerce
This WordPress plugin allows Tezos payments to be made using WooCommerce. It adds a new payment gateway called "Tezos" to the list of available payment gateways in WooCommerce.

When a customer checks out, they will have the option to pay using the Tezos payment gateway. If they choose this option, the plugin will display a button that opens the Temple wallet payment popup. The customer can then sign in to their Temple wallet and complete the payment.

The plugin also includes options for the merchant to customize the title and description of the Tezos payment gateway, as well as enable or disable it. In addition, the plugin includes a field in the plugin's settings for the merchant to enter their Tezos wallet address. This way, the merchant can receive payments from their customers by providing their wallet address.

To use this plugin, the merchant will need to install and activate it on their WordPress website, set up the Tezos payment gateway in the WooCommerce settings, and enter their wallet address in the plugin's settings. They can then start accepting Tezos payments from their customers.