Create Tezos NFT Page Restriction Plugin
To use this plugin, you can add the shortcode [tezos_nft_restriction nft_id="[NFT ID]"] to any page or post that you want to restrict access to. Replace [NFT ID] with the ID of the NFT that is required to view the page.

For example, if you want to restrict access to a page to users who have NFT ID 12345, you can use the shortcode [tezos_nft_restriction nft_id="12345"].

You can also customize the message that is displayed to users who do not have the required NFT by using the message attribute. For example, [tezos_nft_restriction nft_id="12345" message="Sorry, you need to own NFT 12345 to view this page."]

This plugin requires the Temple Wallet plugin to be installed and active on your WordPress website. It uses the temple_wallet_has_nft() function to check if the user has the required NFT in their wallet. If the user does not have the NFT, the plugin displays the message provided in the shortcode. If the user does have the NFT, the plugin does nothing and the user can view the page as normal.